requirements.txt
 PyQt5
 PyNaCl

requests
google-auth
google-auth-httplib2
google-api-python-client
